requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
africastalking>=1.2.6
transformers>=4.40.0
torch>=2.2.0
sentencepiece>=0.2.0          
accelerate>=0.27.0            
soundfile>=0.12.1              
numpy>=1.26.0
cloudinary>=1.36.0
pytest>=8.0.0
pytest-mock>=3.12.0
gunicorn>=21.2.0               
geopy==2.4.1